requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.10
djangorestframework==3.13.0
hmmlearn==0.2.5
joblib==1.0.1
pytz==2021.3
scikit-learn==0.24.1
scipy==1.5.4
sqlparse==0.4.2
threadpoolctl==2.1.0
typing_extensions==4.0.1
